requirements.txt
streamlit 
groq 
python-dotenv
requests
pillow